pyproject.toml
[project]
name = "ai-foundation-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "ipywidgets>=8.0.0",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "langchain-google-genai>=1.0.0",
    "litellm>=1.80.10",
    "mcp>=0.1.0",
    "nest-asyncio>=1.5.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
